Update FocusFileOnSidebar.py
Using sublime's inbuilt methods to check and set sidebar visibility.
The big improvement is that the side bar always gets focused.
The previous version of this plugin suffered from the bug (or feature) that some files, e.g. those not yet saved or not belonging to a folder can not be focused in the side bar. Also a hidden side bar would not be revealed in those cases.